Clarify docstring for EditoState.tr
In the past, I had issues for a while trying to figure out why I couldn't use `state.tr` directly, and had to reassign it. I think the most legible would be `state.tr()` or `state.createTransaction()` where it would be expected to be a different memory reference each time, but adding a little more detail in the comment should help others.